+++ common/src/main/java/com/github/yajatkaul/mega_showdown/gimmick/MegaGimmick.java
@@ -6,6 +6,7 @@ import com.cobblemon.mod.common.api.storage.pc.PCStore;
 import com.cobblemon.mod.common.battles.ShowdownMoveset;
 import com.cobblemon.mod.common.battles.pokemon.BattlePokemon;
 import com.cobblemon.mod.common.pokemon.Pokemon;
+import com.github.yajatkaul.mega_showdown.advancement.AdvancementHelper;
 import com.github.yajatkaul.mega_showdown.codec.Effect;
 import com.github.yajatkaul.mega_showdown.config.MegaShowdownConfig;
 import com.github.yajatkaul.mega_showdown.gimmick.codec.AspectSetCodec;
@@ -91,6 +92,7 @@ public record MegaGimmick(
         if (pokemon.getPersistentData().getBoolean(IS_MEGA_TAG)) {
             unmegaEvolve(pokemon);
         } else {
+            AdvancementHelper.grantAdvancement(pokemon.getOwnerPlayer(), "mega/mega_evolve");
             megaEvolve(pokemon);
         }
     }