+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - fixed tera ghost curse
 - added head locator to msd pokemons
 - fixed mega altaria riding issue
+- client deligate is now a client only mixin as it should have been