+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - fixed liko's pendant
 - fixed shader color bug
 - fixed some advancements not triggering
-- you can right clicked deoxys meteor to change its form
+- you can right click deoxys meteor to change its form
 - shifted `showdown` to `mega_showdown/showdown`
 - now u can use `feature: "mega_evolution"` in species additions directly no need to use species feature assignment
 - Added Tera Particles