+++ CHANGELOG.md
@@ -1,2 +1,3 @@
-# 1.6.2+1.7.2-release
-- 1.7.2 Compatibility
+# 1.6.3+1.7.2-release
+- Added debug mode config to see form change id for battle form
+- Removed unused mixins